Pretty printing overhaul: part I.
Use the pretty printing facilities offered by lib/pp.ml and
lib/util.ml. This avoids the quadratic complexity of concatenating
strings, so should be much faster. In addition, this change affords us
much greater flexibility in the formatting of the output - we can
indent things properly, or not, depending what's appropriate.