Kind declarations and Kind-level rewriting with -coc
If my understanding of the Calculus of Constructions Modulo (chapter
2.8 in Ronan's thesis) is correct, Dedukti should accept declarations
and rewrite rules in Kind when -coc is used. The following file for
example should be accepted:

For example, the following file should be accepted:
```

a : Kind.

def b : Kind.
[] b --> Type.

def c : Kind := Type.

def d := Type.

thm e : Kind := Type.
```

This commit implements this feature.