kernel/rules: Rename partially_typed_rule by untyped_rule
We keep the partially_typed_context type but rename the type for rule
which had a strange meaning from the user point of view.

We also remove the use of unit rule which is redundant with the
partially_typed_rule. Moreover, this type is not exposed to the user
and does not bring new static guarantees where it was used.