kernel/rules: add a ctx field in the type rule_infos
This field is only here to avoid recomputation of the context when the
rule is casted from a rule_infos to an untyped_rule_infos.

This way we can guarantee that the order of variables in the context
is not changed while printing the rule and while it is exported which
may ease the debugging. Moreover, this allows to keep the type
annotations if they were provided by the user