allow more letters in identifiers (#333)
- allow ' as first letter in an identifier (fix #328) 
- allow more letters in identifiers: + * ~ & ^ @ = $ % / < | \ - >

The main modifs are in parsing/lexer.mll and parsing/menhir_parser.mly.
Spaces are added in test files (e.g. writing A -> A instead of A->A).
The error code for tests/KO/nsteps1.dk (`#EVAL[-1]`) is changed in tests/main.ml from lexing error to scoping error because '-' is now allowed in identifiers but the argument of `#EVAL` must be a non-negative integer.